+++ cardinal/gamestate/tick.go
@@ -4,15 +4,14 @@ import (
 	"context"
 	"pkg.world.dev/world-engine/cardinal/codec"
 	"pkg.world.dev/world-engine/cardinal/types"
+	"pkg.world.dev/world-engine/cardinal/types/txpool"
 	"time"
 
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer"
 
+	"github.com/redis/go-redis/v9"
 	"github.com/rotisserie/eris"
 	"pkg.world.dev/world-engine/cardinal/statsd"
-	"pkg.world.dev/world-engine/cardinal/txpool"
-
-	"github.com/redis/go-redis/v9"
 	"pkg.world.dev/world-engine/sign"
 )
