+++ cardinal/router/mocks/router.go
@@ -6,11 +6,11 @@ package mocks
 
 import (
 	context "context"
+	"pkg.world.dev/world-engine/cardinal/types/txpool"
 	reflect "reflect"
 
 	gomock "github.com/golang/mock/gomock"
 	iterator "pkg.world.dev/world-engine/cardinal/router/iterator"
-	txpool "pkg.world.dev/world-engine/cardinal/txpool"
 )
 
 // MockRouter is a mock of Router interface.
